Ignorando os tempos menores
Ignorando os tempos menores

+++ C/bib/infect.h
@@ -2,7 +2,7 @@
 #define INFECT_H
 struct Graph;
 
-void generate_infect(uint16_t N,double p,int seed, int redes,double f,int vacina,bool weight);
+void generate_infect(double N,double p,int seed, int redes,double f,int vacina,bool weight);
 
 
 #endif
